JP-34: feat: also show verify/resend dialog when trying to login with unverified mail
